refactor/feat: process async jobs in child process
